+++ lib/loot.h
@@ -22,7 +22,8 @@ extern char 		 is_loot_stackable(short int id);
 extern void		 pickup_item(struct game *cur_game, struct worldmap *map, struct player *cur_player);
 extern void		 throw_item(struct game *cur_game, struct worldmap *map, struct player *cur_player);
 extern void		 move_cursor(struct game *cur_game, char dir);
-extern void		 move_cursor_click(struct game *cur_game, int x);
 extern void		 swap_item(struct game *cur_game, struct worldmap *map, struct player *cur_player);
+extern SDL_bool		 handle_swap(struct game *cur_game, struct player *cur_player, char start_pos);
+
 
 #endif